solution/pyproject.toml
[project]
name = "solution"
version = "0.1.0"
description = ""
authors = [
    {name = "IvanShumilov24",email = "ivansergshum@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic[email] (>=2.10.5,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "hypothesis (>=6.124.2,<7.0.0)",
    "pydantic-extra-types (>=2.10.2,<3.0.0)",
    "pycountry (>=24.6.1,<25.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
